refactor(capabilities): rename retrieval capabilities and consolidate into automated_retrieval
Renames the three retrieval capabilities and moves the sequential stacker into
the `automated_retrieval` package alongside the random-access one:

- `LoadingTrayRetrieval` (base) -> `AutomatedRetrieval`
- `AutomatedRetrieval` (random access) -> `RandomAccessRetrieval`
- `Stacker` -> `StackerRetrieval`

Both concrete capabilities now live under
`pylabrobot/capabilities/automated_retrieval/`; the old top-level
`loading_tray_retrieval.py` and the `stacker/` package are removed, with
`StackerBackend`/`StackerChatterboxBackend` folded into the package's
`backend.py`/`chatterbox.py`. Backend class names are unchanged, so the
Liconic and Cytomat backends are untouched; their device code now constructs
`RandomAccessRetrieval`. Docs and CHANGELOG updated to match.
