+++ pylabrobot/capabilities/automated_retrieval/__init__.py
@@ -1,2 +1,5 @@
-from .automated_retrieval import AutomatedRetrieval, NoFreeSiteError
-from .backend import AutomatedRetrievalBackend
+from .automated_retrieval import AutomatedRetrieval
+from .backend import AutomatedRetrievalBackend, StackerBackend
+from .chatterbox import AutomatedRetrievalChatterboxBackend, StackerChatterboxBackend
+from .random_access_retrieval import NoFreeSiteError, RandomAccessRetrieval
+from .stacker_retrieval import EmptyStackError, LoadingTrayOccupiedError, StackerRetrieval
