--- pylabrobot/capabilities/stacker/__init__.py
@@ -1,3 +1,0 @@
-from .backend import StackerBackend
-from .chatterbox import StackerChatterboxBackend
-from .stacker import EmptyStackError, LoadingTrayOccupiedError, Stacker
